GIMME THAT TYPESCRIPT
