+++ src/commands/index.ts
@@ -1,0 +1,4 @@
+// Barrel file ot unify exports from the directory
+
+export * from "./ping";
+export * from "./pubSub";
